split main code up more, small media3 changes
